chat-ai-gemini/requirements.txt
shiny
python-dotenv
tokenizers
openai
google.auth
